Fix JSON Table Schema imports
